requirements.txt
telethon
python-decouple
apscheduler